#480 refactored Connection
